GDC return FIFO doesn't need to be 16 bits nor should the function to write it send 16 bits.

+++ src/hardware/vga_pc98_gdc.cpp
@@ -470,7 +470,7 @@ void PC98_GDC_state::flush_fifo_old(void) {
     }
 }
 
-bool PC98_GDC_state::write_rfifo(const uint16_t c) {
+bool PC98_GDC_state::write_rfifo(const uint8_t c) {
     if (rfifo_write >= PC98_GDC_FIFO_SIZE)
         return false;
